refactor: dissolve api/ app into domain apps
