+++ accounts/urls.py
@@ -1,25 +1,17 @@
 from django.urls import path
-from django.contrib.auth.views import (
-    PasswordResetDoneView,
-    PasswordResetConfirmView,
-    PasswordResetCompleteView,
-)
-from .views import register_view, login_view, logout_view, CustomPasswordResetView
 
+from accounts.views import (
+    CookieTokenBlacklistView,
+    CookieTokenObtainPairView,
+    CookieTokenRefreshView,
+    CurrentUserView,
+    RegisterView,
+)
 
 urlpatterns = [
-    path('register/', register_view, name='register'),
-    path('login/', login_view, name='login'),
-    path('logout/', logout_view, name='logout'),
-    path('password_reset/', CustomPasswordResetView.as_view(), name='password_reset'),
-    path('password_reset/done/', PasswordResetDoneView.as_view(
-        template_name='accounts/password_reset/done.html',
-    ), name='password_reset_done'),
-    path('reset/<uidb64>/<token>/', PasswordResetConfirmView.as_view(
-        template_name='accounts/password_reset/confirm.html',
-        success_url='/accounts/reset/done/',
-    ), name='password_reset_confirm'),
-    path('reset/done/', PasswordResetCompleteView.as_view(
-        template_name='accounts/password_reset/complete.html',
-    ), name='password_reset_complete'),
+    path('auth/token/', CookieTokenObtainPairView.as_view(), name='api_token_obtain'),
+    path('auth/token/refresh/', CookieTokenRefreshView.as_view(), name='api_token_refresh'),
+    path('auth/token/blacklist/', CookieTokenBlacklistView.as_view(), name='api_token_blacklist'),
+    path('auth/register/', RegisterView.as_view(), name='api_register'),
+    path('users/me/', CurrentUserView.as_view(), name='api_me'),
 ]
