--- api/pagination.py
@@ -1,7 +1,0 @@
-from rest_framework.pagination import PageNumberPagination
-
-
-class StandardPagination(PageNumberPagination):
-    page_size = 25
-    page_size_query_param = 'page_size'
-    max_page_size = 100
