+++ market/urls.py
@@ -1,14 +1,17 @@
 from django.urls import path
-from . import views_ui, views_api
+
+from market.views import (
+    AssetDetailView,
+    ChartDataView,
+    ExchangeDetailView,
+    ExchangeListView,
+    FxRatesView,
+)
 
 urlpatterns = [
-    # Market overview and browsing
-    path('', views_ui.market_overview_view, name='market_overview'),
-    path('exchange/<str:exchange_code>/', views_ui.exchange_detail_view, name='exchange_detail'),
-    
-    # Asset detail page
-    path('assets/<str:exchange_code>/<str:asset_symbol>/', views_ui.asset_detail_view, name='asset_detail'),
-        
-    # Chart data API endpoints
-    path('data/assets/<str:exchange_code>/<str:asset_symbol>/', views_api.asset_performance_chart_data_view, name='asset_performance_chart_data'),
+    path('market/exchanges/', ExchangeListView.as_view(), name='api_exchanges'),
+    path('market/exchanges/<str:exchange_code>/', ExchangeDetailView.as_view(), name='api_exchange_detail'),
+    path('market/assets/<str:exchange_code>/<str:ticker>/', AssetDetailView.as_view(), name='api_asset_detail'),
+    path('market/data/<str:exchange_code>/<str:ticker>/', ChartDataView.as_view(), name='api_chart_data'),
+    path('wallets/fx-rates/', FxRatesView.as_view(), name='api_fx_rates'),
 ]
